% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-model.r
\name{linear_model}
\alias{linear_model}
\title{Grab coefficients}
\usage{
linear_model(fo, d, contrasts = NULL)
}
\arguments{
\item{lm_obj}{the linear model object created by the lm() function}
}
\description{
This is a function to grab all coefficients of the linear model.
}
